missed a few.  32->64

+++ src/mjolnir/graphvalidator.cc
@@ -46,7 +46,7 @@ struct HGVRestrictionTypes {
 // Get the GraphId of the opposing edge.
 uint32_t GetOpposingEdgeIndex(const GraphId& startnode,
                               DirectedEdge& edge,
-                              uint32_t wayid,
+                              uint64_t wayid,
                               const GraphTile* tile,
                               const GraphTile* end_tile,
                               std::set<uint32_t>& problem_ways,
@@ -135,7 +135,7 @@ uint32_t GetOpposingEdgeIndex(const GraphId& startnode,
       }
 
       bool match = false;
-      uint32_t wayid2 = 0;
+      uint64_t wayid2 = 0;
       if (edge.is_shortcut()) {
         // Shortcut edges - use must match (or both are links)
         if ((directededge->link() && edge.link()) || (directededge->use() == edge.use())) {
@@ -392,7 +392,7 @@ void validate(
         // node. Set the deadend flag and internal flag (if the opposing
         // edge is internal then make sure this edge is as well)
         std::string end_node_iso;
-        uint32_t wayid = tile->edgeinfo(directededge.edgeinfo_offset()).wayid();
+        uint64_t wayid = tile->edgeinfo(directededge.edgeinfo_offset()).wayid();
         uint32_t opp_index =
             GetOpposingEdgeIndex(node, directededge, wayid, tile, endnode_tile, problem_ways,
                                  dupcount, end_node_iso, transit_level);
